Review: cover drop metric with a real-SDK unit test; drop docker live test
Forces a real batch-queue overflow against a local hung listener so CI
exercises the SDK->global-logger->counter chain every run; an SDK message
rename now fails here instead of silently zeroing the metric. Removes the
build-tagged docker live test (its coverage is now unit-side).
